Move SetHubLocation into SettingsPanel folder
