+++ src/web/containers/SettingsPanel/SetHubLocation/SetHubLocation.tsx
@@ -6,10 +6,10 @@ import { Map } from "ol";
 import PointerInteraction from "ol/interaction/Pointer";
 
 // Jaia
-import { JaiaAPI } from "../../utils/jaia-api";
-import { PortalHubStatus } from "../../shared/PortalStatus";
-import { CommandForHub, GeographicCoordinate, HubCommandType } from "../../shared/JAIAProtobuf";
-import { getGeographicCoordinate } from "../../shared/Utilities";
+import { JaiaAPI } from "../../../utils/jaia-api";
+import { PortalHubStatus } from "../../../shared/PortalStatus";
+import { CommandForHub, GeographicCoordinate, HubCommandType } from "../../../shared/JAIAProtobuf";
+import { getGeographicCoordinate } from "../../../shared/Utilities";
 
 // Style
 import Button from "@mui/material/Button";
